workers/requirements.txt
telethon>=1.34.0
flask>=3.0.0
flask-cors>=4.0.0
google-api-python-client>=2.110.0
google-auth>=2.25.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
gunicorn>=21.2.0
openai>=1.0.0
